requirements.txt
kiteconnect==5.0.1
pandas==2.2.2
pyotp==2.9.0
python-dotenv==1.0.1
python_dateutil==2.8.2
Requests==2.32.3
SQLAlchemy==2.0.32
